Fix: Move non-async exports to separate file to comply with 'use server' directive
